requirements.txt
asgiref==3.8.1
Django==5.1.2
django-multiupload==0.6.1
mysqlclient==2.2.4
pillow==10.4.0
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.2
